Update q_remove_head to handle store failed error
Revise q_remove_head by adding strncpy sp to manage storage failures
error message.

Initially, sp is a argument from qtest with the first character set to
'\0' and the remaining characters padded with 'X', sized to
string_length. Subsequently, qtest intends to compare sp with a
parameter 'check' which is arg[1] of qtest rh.

In the case where sp remains unchanged after testing its first character,
implying that q_remove should replace sp with its value for storage.

Refining the function, if the provided character pointer sp is not NULL,
utilize strncpy to copy the value of the removed element to the character
pointer sp, with a maximum size of bufsize - 1 to reserve space for
null-termination. Furthermore, ensure the copied string is null-terminated
by assigning sp[bufsize - 1] to '\0'.

In the event that the copied string stored in sp doesn't match the
argument[1] 'check', an error will be triggered:
"ERROR: Removed value %s != expected value %s",
which will be pushed by the queue test.