Implement deletion of middle node in queue
Implemented a delete the middle node measure by using a fast-slow pointer
technique.

The q_delete_mid function starts by checking for edge cases:
if the queue is empty or if head is NULL, it returns false.

Utilizing two pointers, fast and slow, it traverses the queue. fast
moves twice as fast as slow, allowing efficient identification of the
middle node. Once found, the next node of middle node would be removed
from the queue, and its memory is deallocated.

Signed-off-by: dig84128 <n96121189@gs.ncku.edu.tw>