Enable sorting queue function to support ordering
Utilizing the non-recursive implementation of quicksort as referenced in
"Optimized QuickSort: C Implementation (Non-Recursive)", the `q_sort`
function is adjusted to accommodate both ascending and descending sorting
orders, controlled by the `descend` parameter.

These enhancements aim to provide a more efficient and flexible sorting mechanism for elements in the queue.

Signed-off-by: dig84128 <n96121189@gs.ncku.edu.tw>